Changes handling of named graphs.
We've decided to step away from the single wiki graph, and instead
use a named graph per source of data. This simplifies the handling of
multiple sources of triples.

With regard to the importing of external data, the graph name of the
imported source could use an identifier that does not occur in dokuwiki
page id's. (Or a namespace could be declared for the imported data, and
made unwritable by convention or through de ACL)