Adds ~~NODATA~~ syntax.
This is currently implemented through the $renderer->info array.

I find this method to be less than satisfactory, as the check made
by a data-adding plugin is convoluted (check for existance of
$R->info['data'], and check for value as well).

It does not seem possible to default any member of this array to true
without editting the core.

Using the PARSER_METADATA_RENDER event with BEFORE advice does not
provide any joy either. The flag should not be persistent in my
opinion, and making it persistent is the only way to garuantee
availability during rendering.

Putting the list of flagged pages into a singleton object was also
considered. This was rejected because it wouldn't make for an elegant
solution.